add util.py and fix imports
